:bug: :lollipop: Remove race condition in initSession
The first couple lines of initSession were working with a class variable - one that is set in another method that could possibly be called from another thread. We now take a copy of the reference before we do the checks (including a check for null!), so that our results stay consistent, no matter what other threads might do!